backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy[asyncio]==2.0.44
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.16.5
pydantic==2.11.7
pydantic-settings==2.10.1
python-jose[cryptography]==3.5.0
email-validator==2.3.0
aiosmtplib==4.0.2
httpx==0.28.1
python-dotenv==1.1.1
pytest==8.4.2
pytest-asyncio==1.1.0
